Switching to Parcel for React compilation
