+++ react-js/.gitignore
@@ -19,3 +19,7 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# Parcel
+/.cache
+/dist
