--- react-js/src/App.css
@@ -1,4 +1,0 @@
-body {
-  margin: 0;
-  font: 16px Helvetica, sans-serif;
-}
