+++ react-js/src/LearnMode.css
@@ -1,9 +1,3 @@
-body {
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
-  margin: 0;
-  padding: 0;
-}
-
 .LearnMode {
   padding: 20px;
 }
