+++ react-js/src/LearnMode.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import './LearnMode.css';
 
-class LearnMode extends React.Component {
+export default class LearnMode extends React.Component {
   render() {
     return (
       <div className="LearnMode">
-        <h1>{this.props.terms[0].word}</h1>
-        <form>
-          <input className="LearnMode-input" type="text" />
+        <h1 className="LearnMode-prompt">{this.props.terms[0].word}</h1>
+        <form className="LearnMode-form">
+          <input autoFocus={true} className="LearnMode-input" type="text" />
           <button className="LearnMode-submit" type="submit">
             Submit
           </button>
@@ -16,5 +16,3 @@ class LearnMode extends React.Component {
     );
   }
 }
-
-export default LearnMode;
