--- react-js/src/components/LearnMode.css
@@ -1,27 +1,0 @@
-.LearnMode {
-	padding: 20px;
-}
-
-.LearnMode-input {
-	border-radius: 2px;
-	border: 2px solid #d9dfe2;
-	font: 15px/1.35 "open sans",opensans-webfont,"lucida grande","lucida sans unicode",arial,sans-serif;
-	margin-right: 5px;
-	outline: 0 none;
-	padding: 8px;
-	resize: none;
-	display: inline-block;
-}
-
-.LearnMode-submit {
-	font: 500 14px/30px "lucida grande","lucida sans unicode",arial,sans-serif;
-    background: #1f6fd9;
-    border-radius: 2px;
-    border: 1px solid #1d66c8;
-    color: #fff;
-    display: inline-block;
-    height: 40px;
-    outline: 0 none;
-    padding: 0 20px;
-    text-align: center;
-}
