+++ react-js/src/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Frontend Interview (React JS)</title>
+    <link rel="stylesheet" href="./index.css" />
+  </head>
+  <body>
+    <main id="root"></main>
+    <script src="./index.js"></script>
+  </body>
+</html>
