+++ vanilla-js/src/index.html
@@ -7,12 +7,14 @@
     <link rel="stylesheet" href="./index.css" />
   </head>
   <body>
-    <main class="LearnMode">
-      <h1 class="LearnMode-prompt">Nebraska</h1>
-      <form class="LearnMode-form">
-        <input autofocus class="LearnMode-input" type="text" />
-        <button class="LearnMode-submit" type="submit">Submit</button>
-      </form>
+    <main id="root">
+      <div class="LearnMode">
+        <h1 class="LearnMode-prompt">Nebraska</h1>
+        <form class="LearnMode-form">
+          <input autofocus class="LearnMode-input" type="text" />
+          <button class="LearnMode-submit" type="submit">Submit</button>
+        </form>
+      </div>
     </main>
     <script src="./index.js"></script>
   </body>
