+++ vanilla-js/src/index.js
@@ -1,19 +1,19 @@
 const TERMS = [
-	{
-		id: 1,
-		word: 'Nebraska',
-		definition: 'Lincoln',
-	},
-	{
-		id: 2,
-		word: 'Massachusetts',
-		definition: 'Boston',
-	},
-	{
-		id: 3,
-		word: 'California',
-		definition: 'Sacramento',
-	},
+  {
+    id: 1,
+    word: 'Nebraska',
+    definition: 'Lincoln',
+  },
+  {
+    id: 2,
+    word: 'Massachusetts',
+    definition: 'Boston',
+  },
+  {
+    id: 3,
+    word: 'California',
+    definition: 'Sacramento',
+  },
 ];
 
 // TODO: Write code here
